refactor: extract shared GenericFieldControl from block editor script
Moved to src/js/shared/ so the shortcode builder app can render the
same per-type controls; .eslintrc JSX override widened to the new dirs.
